Consolidate elibc & libc malloc wrappers
